refactor(cfg): Move CFG to behavior module and add basic structure for RemoteExecutionNode
