+++ machine_data_model/behavior/call_method_node.py
@@ -1,16 +1,17 @@
 from typing import Any
 
-from machine_data_model.nodes.composite_method.control_flow_node import (
+from machine_data_model.behavior.control_flow_node import (
     ControlFlowNode,
     resolve_value,
+    LocalExecutionNode,
 )
-from machine_data_model.nodes.composite_method.control_flow_scope import (
+from machine_data_model.behavior.control_flow_scope import (
     ControlFlowScope,
 )
 from machine_data_model.nodes.method_node import AsyncMethodNode
 
 
-class CallMethodNode(ControlFlowNode):
+class CallMethodNode(LocalExecutionNode):
     """
     Represents the call operation of a method in the machine data model. When executed,
     it calls the method with the specified arguments and stores the return values in the scope.
