+++ machine_data_model/builder/data_model_dumper.py
@@ -3,18 +3,18 @@ import os
 import yaml
 
 from machine_data_model.data_model import DataModel
-from machine_data_model.nodes.composite_method.call_method_node import CallMethodNode
+from machine_data_model.behavior.call_method_node import CallMethodNode
 from machine_data_model.nodes.composite_method.composite_method_node import (
     CompositeMethodNode,
 )
-from machine_data_model.nodes.composite_method.control_flow import ControlFlow
-from machine_data_model.nodes.composite_method.read_variable_node import (
+from machine_data_model.behavior.control_flow import ControlFlow
+from machine_data_model.behavior.read_variable_node import (
     ReadVariableNode,
 )
-from machine_data_model.nodes.composite_method.wait_condition_node import (
+from machine_data_model.behavior.wait_condition_node import (
     WaitConditionNode,
 )
-from machine_data_model.nodes.composite_method.write_variable_node import (
+from machine_data_model.behavior.write_variable_node import (
     WriteVariableNode,
 )
 from machine_data_model.nodes.folder_node import FolderNode
@@ -205,7 +205,7 @@ def _composite_method_node_representer(
 
 def _control_flow_graph_representer(
     dumper: yaml.Dumper, node: ControlFlow
-) -> yaml.nodes.MappingNode:
+) -> yaml.nodes.SequenceNode:
     """
     Represent a ControlFlow as a YAML mapping node.
     :param dumper: The YAML dumper instance.
