+++ tests/nodes/composite_method/test_composite_method.py
@@ -7,8 +7,8 @@ from machine_data_model.nodes.composite_method.composite_method_node import (
     CompositeMethodNode,
     SCOPE_ID,
 )
-from machine_data_model.nodes.composite_method.control_flow import ControlFlow
-from machine_data_model.nodes.composite_method.wait_condition_node import (
+from machine_data_model.behavior.control_flow import ControlFlow
+from machine_data_model.behavior.wait_condition_node import (
     WaitConditionNode,
 )
 from machine_data_model.nodes.method_node import AsyncMethodNode
