+++ tests/nodes/composite_method/test_control_flow.py
@@ -1,6 +1,6 @@
 import pytest
 
-from machine_data_model.nodes.composite_method.control_flow_scope import (
+from machine_data_model.behavior.control_flow_scope import (
     ControlFlowScope,
 )
 from machine_data_model.nodes.method_node import AsyncMethodNode
